my-reco-functions/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
numpy==1.26.4
pandas
surprise
faiss-cpu
pyarrow  
